Add success task to publish alias tags

+++ src/prepare.ts
@@ -2,6 +2,22 @@ import * as fs from "fs";
 import execa from "execa";
 import { Context } from "semantic-release";
 
+function getDependencies(branch: any, dev: boolean): Record<string, string> {
+    const dependencies = dev ? branch.devDependencies : branch.dependencies;
+
+    if (!Array.isArray(dependencies)) {
+        return dependencies;
+    }
+
+    const dependencyMap: Record<string, string> = {};
+
+    for (const pkgName of dependencies) {
+        dependencyMap[pkgName] = branch.channel || (branch.main ? "latest" : branch.name);
+    }
+
+    return dependencyMap;
+}
+
 async function updateDependency(context: Context, pkgName: string, pkgTag: string, dev: boolean): Promise<void> {
     const packageJson = JSON.parse(fs.readFileSync("package.json", "utf-8"));
     const dependencies = packageJson[dev ? "devDependencies" : "dependencies"] || {};
@@ -18,20 +34,20 @@ export default async (pluginConfig: any, context: Context): Promise<void> => {
     const branch = (context as any).branch;
 
     if (branch.dependencies) {
-        for (const [pkgName, pkgTag] of Object.entries(branch.dependencies)) {
-            await updateDependency(context, pkgName, pkgTag as string, false);
+        for (const [pkgName, pkgTag] of Object.entries(getDependencies(branch, false))) {
+            await updateDependency(context, pkgName, pkgTag, false);
         }
     }
 
     if (branch.devDependencies) {
-        for (const [pkgName, pkgTag] of Object.entries(branch.devDependencies)) {
-            await updateDependency(context, pkgName, pkgTag as string, true);
+        for (const [pkgName, pkgTag] of Object.entries(getDependencies(branch, true))) {
+            await updateDependency(context, pkgName, pkgTag, true);
         }
     }
 
-    if (fs.existsSync("CHANGELOG.md")) {
+    if (fs.existsSync("CHANGELOG.md") && context.nextRelease != null) {
         const changelogContents: string = fs.readFileSync("CHANGELOG.md", "utf-8");
-        const pkgVersion = context.nextRelease!.version;
+        const pkgVersion = context.nextRelease.version;
 
         if (changelogContents.includes("## Recent Changes") && !changelogContents.includes("## `" + pkgVersion + "`")) {
             fs.writeFileSync("CHANGELOG.md", changelogContents.replace("## Recent Changes", "## `" + pkgVersion + "`"));
